+++ lib/modules/theme/theme_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: theme
-// Updated at: 2026-08-28 12:44:36
-// Commit: style(theme): harmonize light and dark theme palette for theme
+// Updated at: 2026-08-29 19:14:56
+// Commit: fix(theme): fix memory leak by cancelling theme subscription
 
 class themeComponentState {
-  final String commitRef = '3987';
-  final String lastMessage = 'style(theme): harmonize light and dark theme palette for theme';
-  final DateTime timestamp = DateTime.parse('2026-08-28T12:44:36+05:00');
+  final String commitRef = '4042';
+  final String lastMessage = 'fix(theme): fix memory leak by cancelling theme subscription';
+  final DateTime timestamp = DateTime.parse('2026-08-29T19:14:56+05:00');
 }